+++ docs_src/tutorial/build_sdist/vite.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "vite";
+
+// https://vitejs.dev/config/
+export default defineConfig(async () => ({
+  // ...
+  build: {
+    // build dist into `src-tauri/` so that we can include it in Python sdist
+    outDir: "src-tauri/frontend",
+  },
+}));